+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+
+## 2020-06-11
+
+### Added
+
+- COVID-19 Johns Hopkins University global time series data
+
+
 ## 2020-06-10
 
 ### Changed